add assertion test for responder impls
